prototype multi-cloud HDF5 reader
